requirements.txt
Flask==2.2.2
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
google-cloud-error-reporting==1.6.3
google-cloud-firestore==2.7.2
google-cloud-logging==3.2.5
google-cloud-pubsub==2.13.11
google-cloud-tasks==2.10.4
